+++ patzilla-ui/access/epo-ops.js
@@ -198,8 +198,7 @@ OpsExchangeMetadata = Backbone.Model.extend({
         // these carry state, so switching the navigator into a special mode, currently
         reviewmode: false,
 
-        // TODO: Don't have this fixed to OPS
-        datasource: 'ops',
+        datasource: null,
         searchmode: null,
         page_size: 25,
         result_count: null,
@@ -211,7 +210,8 @@ OpsExchangeMetadata = Backbone.Model.extend({
         pagination_pagesize_choices: [25, 50, 75, 100],
         pagination_current_page: 1,
         keywords: [],
-        // TODO: This is also OPS-specific
+
+        // Fixme: This value is specific to OPS.
         maximum_results: 2000,
 
         get_url: function() {
@@ -893,7 +893,7 @@ OpsFulltext = Marionette.Controller.extend({
             .then(function(payload) {
                 if (payload) {
                     var claims = payload['ops:world-patent-data']['ftxt:fulltext-documents']['ftxt:fulltext-document']['claims'];
-                    //console.log('claims', claims);
+                    //console.log('claims', _this.document_number, claims);
 
                     var response = _this.collect_fulltext_items(claims, function(item) { return item['claim']['claim-text']; });
                     deferred.resolve(response, _this.get_datasource_label());
@@ -917,7 +917,7 @@ OpsFulltext = Marionette.Controller.extend({
             .then(function(payload) {
                 if (payload) {
                     var description = payload['ops:world-patent-data']['ftxt:fulltext-documents']['ftxt:fulltext-document']['description'];
-                    //console.log('description', document_number, description);
+                    //console.log('description', _this.document_number, description);
 
                     var response = _this.collect_fulltext_items(description, function(item) { return item['p']; });
                     deferred.resolve(response, _this.get_datasource_label());