+++ CHANGES.rst
@@ -11,6 +11,7 @@ Development
 - [ui] Offer multiple languages when displaying fulltext content
 - [ui] Honor vendor-based data source preference
 - [ui] Optionally expand maximum share time to five years
+- [ui] Fix preferred data source selector
 
 
 2019-05-15 0.167.2